pyproject.toml
[tool.poetry]
name = "wierre-bot"
version = "0.3.0"
description = ""
authors = ["Noah <50159154+kevinshome@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "wierre_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
py-cord = "<2.4.1"
python-dotenv = "^0.21.0"
pygit2 = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
